spapr_pci: Add windows counter
We are going to have more than one DMA window and this is going to
be dynamic. This adds a counter which will be used to get a new DMA
window LIOBN. It is set to 1 by default as window #0 is there by default.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
---
Changes:
v3:
* the counter is int32_t now instead of uint32_t as it might be used
as an array size for migration and VMSTATE_STRUCT_VARRAY works with signed
int