spapr_iommu: Add root memory region
We are going to have multiple DMA windows on at different offsets on
a PCI bus. In order to support a migration, we will need a way to map
just migrated DMA window on PCI to make it available to the guest.

This adds a "root" memory region (UINT64_MAX long) which belongs to
a TCE table and which is mapped onto PCI bus with overlapping enabled as
there will more than just one window at the time.

The actual TCE table is a subregion of the "root" at @bus_offset.
This way migration can map the actual TCE table at correct location.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>