spapr_pci: Rework finish_realize()
This renames and reworks finish_realize() which used to finalize DMA
setup assuming that it will not change later.

The new callback can create more that one window and supports various
parameters such as page and windows sizes.

This is a mechanical change so no change in behaviour is expected.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>