spapr_pci: Rework reset to reset DMA configuration
On a system reset, A SPAPR guest expects the DMA configuration to be
reset too. Without DDW, this does not require any changes to QEMU but
when we enable DDW, we will have to remove all DMA windows and create
one default window at zero offset on a PCI bus, 4K pages.

This reworks SPAPR PHB code to disable the existing DMA window on reset
and then configure and enable the default window.
Without DDW that means that the same window will be disabled and then
enabled with no other change in behaviour.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>

Conflicts:
	hw/ppc/spapr_pci_vfio.c