vfio: spapr: Add DMA memory registering
This makes use of the new "memory registering" feature. The idea is
to provide the guest ability to notify the host kernel about pages which
are going to be used for DMA. Having this information, the host kernel
can pin them all, do locked pages accounting and not spent time on
doing that in real time with possible failures which cannot be handled
nicely in some cases.

This adds a memory listener which notifies a VFIO container about memory
which needs to be pinned/unpinned. VFIO MMIO regions (i.e. "skip dump"
regions) are skipped.

The feature is only enabled for SPAPR IOMMU. The host kernel changes
are required.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
---
Changes:
v5:
* simplified the patch
* added trace points
* added round_up() for the size
* added missing memory_region_unref()