spapr_pci/spapr_pci_vfio: Support Dynamic DMA Windows (DDW)
This implements DDW for emulated and VFIO PHB.

This reserves RTAS token numbers for DDW calls.

This adds callbacks definitions which PHB needs to implement in order to
support dynamic DMA windows (DDW).

This implements ddw_query() callback which returns DDW support
constraints. This is not a static set of properties as for VFIO these
are read from the host kernel.

This extends init_dma_window() callbacks to support a second DMA window
which is going to contain mapping of the entire guest RAM in most cases.

This changes the TCE table migration descriptor to support dynamic
tables as from now on, PHB will create as many stub TCE table objects
as PHB can possibly support but not all of them might be initialized at
the time of migration because DDW might or might not be requested by
the guest.

The "ddw" property is enabled by default on a PHB but for compatibility
pseries-2.2 machine disables it.

This implements DDW for VFIO. Host kernel support is required for this.
This adds a "levels" property to VFIO PHB to control the number of levels
in the actual TCE table allocated by the host kernel, 0 is the default
value to tell QEMU to calculate the correct value. Current hardware
supports up to 5 levels.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
---
Changes:
v5:
* total rework
* enabled for machines >2.2
* fixed migration

v4:
* reset handler is back in generalized form

v3:
* removed reset
* windows_num is now 1 or bigger rather than 0-based value and it is only
changed in PHB code, not in RTAS
* added page mask check in create()
* added SPAPR_PCI_DDW_MAX_WINDOWS to track how many windows are already
created

v2:
* tested on hacked emulated E1000
* implemented DDW reset on the PHB reset
* spapr_pci_ddw_remove/spapr_pci_ddw_reset are public for reuse by VFIO