vfio: Enable in-kernel acceleration via VFIO KVM device
TCE hypercalls (H_PUT_TCE, H_PUT_TCE_INDIRECT, H_STUFF_TCE) use a logical bus
number (LIOBN) to identify which TCE table the request is addressed to.
However VFIO kernel driver operates with IOMMU group IDs and has no idea
about which LIOBN corresponds to which group. If the host kernel supports
in-kernel acceleration for TCE calls, we have to provide the LIOBN to IOMMU
mapping information.

This makes use of a VFIO KVM device's
KVM_DEV_VFIO_GROUP_SET_SPAPR_TCE_LIOBN attribute to set the link
between LIOBN and IOMMU group. This falls back to user-space TCE table
emulation if KVM failed to create the link.

The vfio_container_spapr_set_liobn() helper is placed
in vfio.c as it needs to convert the address space to a group fd and
this association is not available outside VFIO code.

This adds vfio-pci handling when KVM acceleration is available.

This removes vfio_accel flag from PHB as it is used in the same context
where it is set.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>