spapr_pci: Do not try removing absent huge window
The PHB reset handler resets all DMA windows unconditionally which
may result in failing ioctls to the VFIO IOMMU host driver if there was
no window at the specific bus address. For example, if any VFIO device
is attached, before the guest even started, there is one default DMA
window (which is ok to remove) but there is no huge window so an error
message is printed:
qemu-system-ppc64: vfio: failed to ioctl 20 to container: ret=-1, Invalid argument

This makes ioctl to VFIO conditional: remove the first window always
(as it always exists at the first boot and after each reset)
and remove the second window only if it was enabled.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>