vfio: Enable in-kernel acceleration via VFIO KVM device
TCE hypercalls (H_PUT_TCE, H_PUT_TCE_INDIRECT, H_STUFF_TCE) use
a logical bus number (LIOBN) to identify which TCE table the request
is addressed to. However VFIO kernel driver operates with IOMMU group
IDs and has no idea about which LIOBN corresponds to which group.
If the host kernel supports in-kernel acceleration for TCE calls,
we have to provide the LIOBN to IOMMU mapping information.

This makes use of a VFIO KVM device's
KVM_DEV_VFIO_GROUP_SET_SPAPR_TCE_LIOBN attribute to set the link
between LIOBN and IOMMU group.

This adds vfio-pci handling when KVM acceleration is available.

This adds a get_kvm_id() callback to IOMMU MR to let the caller know
if KVM acceleration is possible and if so, what the table ID in KVM.
The callback returns -1 if there is no acceleration available.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>