spapr_iommu: Realloc guest visible TCE table when hot(un)plugging vfio-pci
This replaces g_malloc() with spapr_tce_alloc_table() as this is
the standard way of allocating tables and this allows moving the table
moving back to KVM at no cost.

This enables transition back to KVM; however at the moment the only
caller of spapr_tce_set_need_vfio() calls it with need_vfio==false so
spapr_tce_alloc_table() always allocates a table in the userspace and
return fd==-1.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>