exec: Add helper to return smallest page size from memory backends
Relying on @mempath in order to know whether the guest is backed by
huges pages is not enough as there is a "memory-file-backend" class
which also got "mem-path".

This implements getrampagesize() helper which returns the smallest
page size. The implementation is borrowed from target-ppc/kvm.c which
will be removed later.

While we are here, remove the HUGETLBFS_MAGIC definition as it is not
used anyway.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>