spapr_pci: Get rid of dma_loibn
We are going to have 2 DMA windows which LIOBNs are calculated from
the PHB index and the window number using the SPAPR_PCI_LIOBN macro
so there is no actual use for dma_liobn.

This replaces dma_liobn with SPAPR_PCI_LIOBN. This marks it as unused
in the migration stream. This renames dma_liobn to _dma_liobn as we have
to keep the property for the CLI compatibility and we need a storage
for it, although it has never really been used.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
---
Changes:
v15:
* new to the series