spapr: Create ibm,gpu and ibm,npu cross links in the device tree
These are required by the recent NVIDIA driver in order to support
NVLink2; NVLink v1 did not require that.

This adds an additional PHB tree traversal to collect GPU and NPUs
pointers; the FDT renderer uses these to create the actual properties.

For simplicity this assumes that only 1 GPU and corresponding NVLink2
bridges are attached to a given PHB.

SLOF needs an update for consistency, even though it works today
with phandles generated by QEMU.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>