spapr-pci: Allow configuring maximum DMA window size
At the moment we limit the size of a DMA window by the max RAM size;
however when we start supporting memory devices which appear behind
the max ram size boundary, we won't be able to map this new memory for
direct DMA mapping.

This allows relaxing the upper limit via cli. This is going to be used
later to extend the DMA window size to cover NVIDIA V100 GPU RAM.

The new "max-dma-window" parameter takes a number of TCEs as this is what
the RTAS "ibm,query-pe-dma-window" handler returns. Zero means fallback
to the default limit of max RAM size.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>