spapr_pci: Make maximum supported DMA windows a property
At the moment we always advertise 2 DMA windows support and since
32bit window is always enabled by default, the guest normally does not
touch it and simply creates a bigger one to map the entire guest memory.

However the only hardware platform supporting the second window provides
that window at 1<<59 offset and cannot be used by devices capable of less
bits for PCI bus access.

This replaces the hard coded limit of 2 windows with a PHB property which
the guest can use it as the hint to remove the default 32bit window and
create a 64bit one but starting at the zero offset this time.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>