vfio/spapr: Fix indirect levels calculation
The current code assumes that we can address more bits on a PCI bus
for DMA than we really do. Limit to the known maximum of 55 bits and
try less levels if even that is too much; worth mentioning that
the IOMMU pagesize also defines the size of indirect tables as well.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>