spapr/vfio: Map GPU RAM and advertise to the guest
The NVIDIA V100 GPU comes with some on-board RAM which is mapped into
the host memory space and (when trained) accessible as normal RAM.
The VFIO-PCI driver implements a special region which provides mapping
access for this area so we can map it to QEMU, register as a KVM memory
slot and therefore provide direct access for the guest.

This adds a quick to VFIO to map the memory and create an MR; the new MR
is stored in a GPU as a QOM link. The sPAPR PCI uses this to get the MR
and map it to the system address space at location selected in
sPAPRPHBState::nv2_gpa.

The FDT renderer puts a memory@ node for every GPU RAM chunk with
a size field of the "linux,usable-memory" property set to zero to prevent
the guest from using it until the NVIDIA driver in guest trains the memory
link.

This adds "ibm,device-tgt-addr" (the encoded guest physical address of
the GPU RAM), "ibm,nvlink-speed" (some magic value), "memory-region"
(a phandle of the new memory node) to every NPU node as the NVIDIA
driver relies on these being present.

This puts new memory nodes in a separate NUMA node to replicate the host
system setup as close as possible (the NVIDIA driver relies on this too).

Known issues:

1. Since this is an RFC, this simplifies the design by allowing just one
GPU+2xNPU group per vPHB (not enforced though).

2. The unpatched guest kernel will only use the 32bit DMA window as
the V100 GPU only supports 46 bits (or similar) for DMA and therefore
cannot use the default offset we have for 64bit DMA windows which is 1<<59.

3. We need to provide the guest with enough information about what
the maximum RAM size is. The unpatched guest looks at the last memory
block end and the last block in the DR conf list. The linux kernel does
not add memory blocks if size of "linux,usable-memory" is null
(and hacking that is quite ugly); DR subsystem does not seem to support
sparse memory at the moment so trying "-m XX,maxram=48000T" results
in enourmous FDT blobs (hundreds of MB) so the guest cannot boot.
The current approach is to walk through memory@ nodes in the guest and
calculate the maximum.
Note that a 64bit window of every vPHB must cover the entire guest
RAM including GPU RAM blocks from all GPUs passed through, not just those
which are attached to a specific PHB, because the guest uses 64bit windows
for direct access of everything everywhere.

Sample cmdline:
pbuild/qemu-aikrhel74alt-ppc64/ppc64-softmmu/qemu-system-ppc64 \
-nodefaults \
-chardev stdio,id=STDIO0,signal=off,mux=on \
-device spapr-vty,id=svty0,reg=0x71000100,chardev=STDIO0 \
-mon id=MON0,chardev=STDIO0,mode=readline -nographic -vga none \
-enable-kvm \
-device nec-usb-xhci,id=nec-usb-xhci0 -m 16G \
-chardev socket,id=SOCKET0,server,nowait,host=localhost,port=40000 \
-mon chardev=SOCKET0,mode=control \
-smp 2 \
-netdev "user,id=USER0,hostfwd=tcp::2222-:22" \
-device "virtio-net-pci,id=vnet0,mac=C0:41:49:4b:00:00,netdev=USER0" \
-device \
spapr-pci-host-bridge,id=phb1,index=1,ddw-windows=1,gpa=0x240000000000,dma64_win_addr=0,max-dma-window=0xffffffff \
-device \
"vfio-pci,id=vfio0035_03_00_0,host=0035:03:00.0,bus=phb1.0,addr=0.0" \
-device "vfio-pci,id=vfio0007_00_00_0,host=0007:00:00.0" \
-device "vfio-pci,id=vfio0007_00_00_1,host=0007:00:00.1" \
-kernel ./vmldbg2 -append "root=/dev/sda2 debug loglevel=7" \
u1710-64le-nvidia-396.15.qcow2 -snapshot -global \
spapr-pci-host-bridge.pgsz=0x1011000 -global \
spapr-pci-vfio-host-bridge.pgsz=0x1011000 \
-machine pseries \
-L /home/aik/t/qemu-ppc64-bios/ \
-trace events=qemu_trace_events -d guest_errors

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>